remove MeasurementPair model MeasurementPair objects are not needed in the database and take a long time to insert. Removing them should speed up processing times.
