+++ pyproject.toml
@@ -36,6 +36,7 @@ gevent = { version = "^21.1.2", optional = true }
 gunicorn = { version = "^20.0.4", optional = true }
 forced-phot = { git = "https://github.com/askap-vast/forced_phot.git" }
 strictyaml = "^1.3.2"
+django-stubs-ext = "^0.3.1"
 
 [tool.poetry.dev-dependencies]
 mkdocs-material = "^7.1.4"
@@ -50,16 +51,22 @@ pip-tools = "^5.5.0"
 tabulate = "^0.8.9"
 django-debug-toolbar = "^3.2"
 flake8 = "^3.9.0"
-mypy = "^0.812"
+mypy = "^0.910"
 black = "^20.8b1"
 django-extensions = "^3.1.1"
 rope = "^0.18.0"
-django-stubs = "^1.7.0"
+django-stubs = "^1.9.0"
 lightgallery = "^0.5"
 mkdocs-git-revision-date-localized-plugin = "^0.9"
 
 [tool.poetry.extras]
 prod = ["gevent", "gunicorn"]
+
+[tool.mypy]
+plugins = ["mypy_django_plugin.main"]
+
+[tool.django-stubs]
+django_settings_module = "webinterface.settings"
 
 [build-system]
 requires = ["poetry-core>=1.0.0"]
