+++ vast_pipeline/models.py
@@ -1,3 +1,7 @@
+from dataclasses import dataclass
+from itertools import combinations
+from typing import List
+
 from django.db import models
 from django.core.validators import RegexValidator
 from django.contrib.auth.models import User
@@ -6,6 +10,19 @@ from django.contrib.contenttypes.models import ContentType
 from django.templatetags.static import static
 from social_django.models import UserSocialAuth
 from tagulous.models import TagField
+
+from vast_pipeline.pipeline.pairs import calculate_vs_metric, calculate_m_metric
+
+
+@dataclass
+class MeasurementPair:
+    source_id: int
+    measurement_a_id: int
+    measurement_b_id: int
+    vs_peak: float
+    m_peak: float
+    vs_int: float
+    m_int: float
 
 
 class Comment(models.Model):
@@ -226,6 +243,306 @@ class SourceQuerySet(models.QuerySet):
         )
 
 
+class Image(CommentableModel):
+    """An image is a 2D radio image from a cube"""
+    band = models.ForeignKey(Band, on_delete=models.CASCADE)
+    run = models.ManyToManyField(Run)
+    skyreg = models.ForeignKey(SkyRegion, on_delete=models.CASCADE)
+
+    measurements_path = models.FilePathField(
+        max_length=200,
+        db_column='meas_path',
+        help_text=(
+            'the path to the measurements parquet that belongs to this image'
+        )
+    )
+    POLARISATION_CHOICES = [
+        ('I', 'I'),
+        ('XX', 'XX'),
+        ('YY', 'YY'),
+        ('Q', 'Q'),
+        ('U', 'U'),
+        ('V', 'V'),
+    ]
+    polarisation = models.CharField(
+        max_length=2,
+        choices=POLARISATION_CHOICES,
+        help_text='Polarisation of the image one of I,XX,YY,Q,U,V.'
+    )
+    name = models.CharField(
+        unique=True,
+        max_length=200,
+        help_text='Name of the image.'
+    )
+    path = models.FilePathField(
+        max_length=500,
+        help_text='Path to the file containing the image.'
+    )
+    noise_path = models.FilePathField(
+        max_length=300,
+        blank=True,
+        default='',
+        help_text='Path to the file containing the RMS image.'
+    )
+    background_path = models.FilePathField(
+        max_length=300,
+        blank=True,
+        default='',
+        help_text='Path to the file containing the background image.'
+    )
+
+    datetime = models.DateTimeField(
+        help_text='Date/time of observation or epoch.'
+    )
+    jd = models.FloatField(
+        help_text='Julian date of the observation (days).'
+    )
+    duration = models.FloatField(
+        default=0.,
+        help_text='Duration of the observation.'
+    )
+
+    ra = models.FloatField(
+        help_text='RA of the image centre (Deg).'
+    )
+    dec = models.FloatField(
+        help_text='DEC of the image centre (Deg).'
+    )
+    fov_bmaj = models.FloatField(
+        help_text='Field of view major axis (Deg).'
+    )  # Major (Dec) radius of image (degrees)
+    fov_bmin = models.FloatField(
+        help_text='Field of view minor axis (Deg).'
+    )  # Minor (RA) radius of image (degrees)
+    physical_bmaj = models.FloatField(
+        help_text='The actual size of the image major axis (Deg).'
+    )  # Major (Dec) radius of image (degrees)
+    physical_bmin = models.FloatField(
+        help_text='The actual size of the image minor axis (Deg).'
+    )  # Minor (RA) radius of image (degrees)
+    radius_pixels = models.FloatField(
+        help_text='Radius of the useable region of the image (pixels).'
+    )
+
+    beam_bmaj = models.FloatField(
+        help_text='Major axis of image restoring beam (Deg).'
+    )
+    beam_bmin = models.FloatField(
+        help_text='Minor axis of image restoring beam (Deg).'
+    )
+    beam_bpa = models.FloatField(
+        help_text='Beam position angle (Deg).'
+    )
+    rms_median = models.FloatField(
+        help_text='Background average RMS from the provided RMS map (mJy).'
+    )
+    rms_min = models.FloatField(
+        help_text='Background minimum RMS from the provided RMS map (mJy).'
+    )
+    rms_max = models.FloatField(
+        help_text='Background maximum RMS from the provided RMS map (mJy).'
+    )
+
+    class Meta:
+        ordering = ['datetime']
+
+    def __str__(self):
+        return self.name
+
+
+class MeasurementQuerySet(models.QuerySet):
+
+    def cone_search(
+        self, ra: float, dec: float, radius_deg: float
+    ) -> models.QuerySet:
+        """
+        Return all the Sources withing radius_deg of (ra,dec).
+        Returns a QuerySet of Sources, ordered by distance from
+        (ra,dec) ascending.
+
+        Args:
+            ra: The right ascension value of the cone search central
+                coordinate.
+            dec: The declination value of the cone search central coordinate.
+            radius_deg: The radius over which to perform the cone search.
+
+        Returns:
+            Measurements found withing the cone search area.
+        """
+        return (
+            self.extra(
+                select={
+                    "distance": "q3c_dist(ra, dec, %s, %s) * 3600"
+                },
+                select_params=[ra, dec],
+                where=["q3c_radial_query(ra, dec, %s, %s, %s)"],
+                params=[ra, dec, radius_deg],
+            )
+            .order_by("distance")
+        )
+
+
+class Measurement(CommentableModel):
+    """
+    A Measurement is an object in the sky that has been detected at least once.
+    Essentially a source single measurement in time.
+    """
+    image = models.ForeignKey(
+        Image,
+        null=True,
+        on_delete=models.CASCADE
+    )  # first image seen in
+    source = models.ManyToManyField(
+        'Source',
+        through='Association',
+        through_fields=('meas', 'source')
+    )
+
+    name = models.CharField(max_length=64)
+
+    ra = models.FloatField(help_text='RA of the source (Deg).')  # degrees
+    ra_err = models.FloatField(
+        help_text='RA error of the source (Deg).'
+    )
+    dec = models.FloatField(help_text='DEC of the source (Deg).')  # degrees
+    dec_err = models.FloatField(
+        help_text='DEC error of the source (Deg).'
+    )
+
+    bmaj = models.FloatField(
+        help_text=(
+            'The major axis of the Gaussian fit to the source (Deg).'
+        )
+    )
+    err_bmaj = models.FloatField(help_text='Error major axis (Deg).')
+    bmin = models.FloatField(
+        help_text=(
+            'The minor axis of the Gaussian fit to the source (Deg).'
+        )
+    )
+    err_bmin = models.FloatField(help_text='Error minor axis (Deg).')
+    pa = models.FloatField(
+        help_text=(
+            'Position angle of Gaussian fit east of north to bmaj '
+            '(Deg).'
+        )
+    )
+    err_pa = models.FloatField(help_text='Error position angle (Deg).')
+
+    # supplied by user via config
+    ew_sys_err = models.FloatField(
+        help_text='Systematic error in east-west (RA) direction (Deg).'
+    )
+    # supplied by user via config
+    ns_sys_err = models.FloatField(
+        help_text='Systematic error in north-south (dec) direction (Deg).'
+    )
+
+    # estimate of maximum error radius (from ra_err and dec_err)
+    # Used in advanced association.
+    error_radius = models.FloatField(
+        help_text=(
+            'Estimate of maximum error radius using ra_err'
+            ' and dec_err (Deg).'
+        )
+    )
+
+    # quadratic sum of error_radius and ew_sys_err
+    uncertainty_ew = models.FloatField(
+        help_text=(
+            'Total east-west (RA) uncertainty, quadratic sum of'
+            ' error_radius and ew_sys_err (Deg).'
+        )
+    )
+    # quadratic sum of error_radius and ns_sys_err
+    uncertainty_ns = models.FloatField(
+        help_text=(
+            'Total north-south (Dec) uncertainty, quadratic sum of '
+            'error_radius and ns_sys_err (Deg).'
+        )
+    )
+
+    flux_int = models.FloatField()  # mJy/beam
+    flux_int_err = models.FloatField()  # mJy/beam
+    flux_int_isl_ratio = models.FloatField(
+        help_text=(
+            'Ratio of the component integrated flux to the total'
+            ' island integrated flux.'
+        )
+    )
+    flux_peak = models.FloatField()  # mJy/beam
+    flux_peak_err = models.FloatField()  # mJy/beam
+    flux_peak_isl_ratio = models.FloatField(
+        help_text=(
+            'Ratio of the component peak flux to the total'
+            ' island peak flux.'
+        )
+    )
+    chi_squared_fit = models.FloatField(
+        db_column='chi2_fit',
+        help_text='Chi-squared of the Guassian fit to the source.'
+    )
+    spectral_index = models.FloatField(
+        db_column='spectr_idx',
+        help_text='In-band Selavy spectral index.'
+    )
+    spectral_index_from_TT = models.BooleanField(
+        default=False,
+        db_column='spectr_idx_tt',
+        help_text=(
+            'True/False if the spectral index came from the taylor '
+            'term.'
+        )
+    )
+
+    local_rms = models.FloatField(
+        help_text='Local rms in mJy from Selavy.'
+    )  # mJy/beam
+
+    snr = models.FloatField(
+        help_text='Signal-to-noise ratio of the measurement.'
+    )
+
+    flag_c4 = models.BooleanField(
+        default=False,
+        help_text='Fit flag from Selavy.'
+    )
+
+    compactness = models.FloatField(
+        help_text='Int flux over peak flux.'
+    )
+
+    has_siblings = models.BooleanField(
+        default=False,
+        help_text='True if the fit comes from an island that has more than 1 component.'
+    )
+    component_id = models.CharField(
+        max_length=64,
+        help_text=(
+            'The ID of the component from which the source comes from.'
+        )
+    )
+    island_id = models.CharField(
+        max_length=64,
+        help_text=(
+            'The ID of the island from which the source comes from.'
+        )
+    )
+
+    forced = models.BooleanField(
+        default=False,
+        help_text='True: the measurement is forced extracted.'
+    )
+
+    objects = MeasurementQuerySet.as_manager()
+
+    class Meta:
+        ordering = ['ra']
+
+    def __str__(self):
+        return self.name
+
+
 class Source(CommentableModel):
     run = models.ForeignKey(Run, on_delete=models.CASCADE, null=True,)
     related = models.ManyToManyField(
@@ -351,7 +668,6 @@ class Source(CommentableModel):
         )
     )
 
-
     # total metrics to report in UI
     n_meas = models.IntegerField(
         help_text='total measurements of the source'
@@ -374,326 +690,30 @@ class Source(CommentableModel):
     def __str__(self):
         return self.name
 
-
-class RelatedSource(models.Model):
-    '''
-    Association table for the many to many Source relationship with itself
-    Django doc https://docs.djangoproject.com/en/3.1/ref/models/fields/#django.db.models.ManyToManyField.through
-    '''
-    from_source = models.ForeignKey(Source, on_delete=models.CASCADE)
-    to_source = models.ForeignKey(
-        Source,
-        on_delete=models.CASCADE,
-        related_name='related_sources'
-    )
-
-    class Meta:
-        constraints = [
-            models.UniqueConstraint(
-                name='%(app_label)s_%(class)s_unique_pair',
-                fields=['from_source', 'to_source']
+    def get_measurement_pairs(self) -> List[MeasurementPair]:
+        measurements = (
+            Measurement.objects.filter(source=self)
+            .select_related("image")
+            .order_by("image__datetime")
+        )
+        measurement_pairs: List[MeasurementPair] = []
+        for meas_a, meas_b in combinations(measurements, 2):
+            # ensure the measurements are in time order
+            if meas_a.image.datetime > meas_b.image.datetime:
+                meas_a, meas_b = meas_b, meas_a
+            # calculate metrics
+            vs_peak = calculate_vs_metric(
+                meas_a.flux_peak, meas_b.flux_peak, meas_a.flux_peak_err, meas_b.flux_peak_err
             )
-        ]
-
-
-class Image(CommentableModel):
-    """An image is a 2D radio image from a cube"""
-    band = models.ForeignKey(Band, on_delete=models.CASCADE)
-    run = models.ManyToManyField(Run)
-    skyreg = models.ForeignKey(SkyRegion, on_delete=models.CASCADE)
-
-    measurements_path = models.FilePathField(
-        max_length=200,
-        db_column='meas_path',
-        help_text=(
-            'the path to the measurements parquet that belongs to this image'
-        )
-    )
-    POLARISATION_CHOICES = [
-        ('I', 'I'),
-        ('XX', 'XX'),
-        ('YY', 'YY'),
-        ('Q', 'Q'),
-        ('U', 'U'),
-        ('V', 'V'),
-    ]
-    polarisation = models.CharField(
-        max_length=2,
-        choices=POLARISATION_CHOICES,
-        help_text='Polarisation of the image one of I,XX,YY,Q,U,V.'
-    )
-    name = models.CharField(
-        unique=True,
-        max_length=200,
-        help_text='Name of the image.'
-    )
-    path = models.FilePathField(
-        max_length=500,
-        help_text='Path to the file containing the image.'
-    )
-    noise_path = models.FilePathField(
-        max_length=300,
-        blank=True,
-        default='',
-        help_text='Path to the file containing the RMS image.'
-    )
-    background_path = models.FilePathField(
-        max_length=300,
-        blank=True,
-        default='',
-        help_text='Path to the file containing the background image.'
-    )
-
-    datetime = models.DateTimeField(
-        help_text='Date/time of observation or epoch.'
-    )
-    jd = models.FloatField(
-        help_text='Julian date of the observation (days).'
-    )
-    duration =  models.FloatField(
-        default=0.,
-        help_text='Duration of the observation.'
-    )
-
-    ra = models.FloatField(
-        help_text='RA of the image centre (Deg).'
-    )
-    dec = models.FloatField(
-        help_text='DEC of the image centre (Deg).'
-    )
-    fov_bmaj = models.FloatField(
-        help_text='Field of view major axis (Deg).'
-    )# Major (Dec) radius of image (degrees)
-    fov_bmin = models.FloatField(
-        help_text='Field of view minor axis (Deg).'
-    )# Minor (RA) radius of image (degrees)
-    physical_bmaj = models.FloatField(
-        help_text='The actual size of the image major axis (Deg).'
-    )# Major (Dec) radius of image (degrees)
-    physical_bmin = models.FloatField(
-        help_text='The actual size of the image minor axis (Deg).'
-    )# Minor (RA) radius of image (degrees)
-    radius_pixels = models.FloatField(
-        help_text='Radius of the useable region of the image (pixels).'
-    )
-
-    beam_bmaj = models.FloatField(
-        help_text='Major axis of image restoring beam (Deg).'
-    )
-    beam_bmin = models.FloatField(
-        help_text='Minor axis of image restoring beam (Deg).'
-    )
-    beam_bpa = models.FloatField(
-        help_text='Beam position angle (Deg).'
-    )
-    rms_median = models.FloatField(
-        help_text='Background average RMS from the provided RMS map (mJy).'
-    )
-    rms_min = models.FloatField(
-        help_text='Background minimum RMS from the provided RMS map (mJy).'
-    )
-    rms_max = models.FloatField(
-        help_text='Background maximum RMS from the provided RMS map (mJy).'
-    )
-
-    class Meta:
-        ordering = ['datetime']
-
-    def __str__(self):
-        return self.name
-
-
-class MeasurementQuerySet(models.QuerySet):
-
-    def cone_search(
-        self, ra: float, dec: float, radius_deg: float
-    ) -> models.QuerySet:
-        """
-        Return all the Sources withing radius_deg of (ra,dec).
-        Returns a QuerySet of Sources, ordered by distance from
-        (ra,dec) ascending.
-
-        Args:
-            ra: The right ascension value of the cone search central
-                coordinate.
-            dec: The declination value of the cone search central coordinate.
-            radius_deg: The radius over which to perform the cone search.
-
-        Returns:
-            Measurements found withing the cone search area.
-        """
-        return (
-            self.extra(
-                select={
-                    "distance": "q3c_dist(ra, dec, %s, %s) * 3600"
-                },
-                select_params=[ra, dec],
-                where=["q3c_radial_query(ra, dec, %s, %s, %s)"],
-                params=[ra, dec, radius_deg],
+            m_int = calculate_m_metric(meas_a.flux_int, meas_b.flux_int)
+            vs_int = calculate_vs_metric(
+                meas_a.flux_int, meas_b.flux_int, meas_a.flux_int_err, meas_b.flux_int_err
             )
-            .order_by("distance")
-        )
-
-
-class Measurement(CommentableModel):
-    """
-    A Measurement is an object in the sky that has been detected at least once.
-    Essentially a source single measurement in time.
-    """
-    image = models.ForeignKey(
-        Image,
-        null=True,
-        on_delete=models.CASCADE
-    )# first image seen in
-    source = models.ManyToManyField(
-        Source,
-        through='Association',
-        through_fields=('meas', 'source')
-    )
-
-    name = models.CharField(max_length=64)
-
-    ra = models.FloatField(help_text='RA of the source (Deg).')# degrees
-    ra_err = models.FloatField(
-        help_text='RA error of the source (Deg).'
-    )
-    dec = models.FloatField(help_text='DEC of the source (Deg).')# degrees
-    dec_err = models.FloatField(
-        help_text='DEC error of the source (Deg).'
-    )
-
-    bmaj = models.FloatField(
-        help_text=(
-            'The major axis of the Gaussian fit to the source (Deg).'
-        )
-    )
-    err_bmaj = models.FloatField(help_text='Error major axis (Deg).')
-    bmin = models.FloatField(
-        help_text=(
-            'The minor axis of the Gaussian fit to the source (Deg).'
-        )
-    )
-    err_bmin = models.FloatField(help_text='Error minor axis (Deg).')
-    pa = models.FloatField(
-        help_text=(
-            'Position angle of Gaussian fit east of north to bmaj '
-            '(Deg).'
-        )
-    )
-    err_pa = models.FloatField(help_text='Error position angle (Deg).')
-
-    # supplied by user via config
-    ew_sys_err = models.FloatField(
-        help_text='Systematic error in east-west (RA) direction (Deg).'
-    )
-    # supplied by user via config
-    ns_sys_err = models.FloatField(
-        help_text='Systematic error in north-south (dec) direction (Deg).'
-    )
-
-    # estimate of maximum error radius (from ra_err and dec_err)
-    # Used in advanced association.
-    error_radius = models.FloatField(
-        help_text=(
-            'Estimate of maximum error radius using ra_err'
-            ' and dec_err (Deg).'
-        )
-    )
-
-    # quadratic sum of error_radius and ew_sys_err
-    uncertainty_ew = models.FloatField(
-        help_text=(
-            'Total east-west (RA) uncertainty, quadratic sum of'
-            ' error_radius and ew_sys_err (Deg).'
-        )
-    )
-     # quadratic sum of error_radius and ns_sys_err
-    uncertainty_ns = models.FloatField(
-        help_text=(
-            'Total north-south (Dec) uncertainty, quadratic sum of '
-            'error_radius and ns_sys_err (Deg).'
-        )
-    )
-
-    flux_int = models.FloatField()# mJy/beam
-    flux_int_err = models.FloatField()# mJy/beam
-    flux_int_isl_ratio = models.FloatField(
-        help_text=(
-            'Ratio of the component integrated flux to the total'
-            ' island integrated flux.'
-        )
-    )
-    flux_peak = models.FloatField()# mJy/beam
-    flux_peak_err = models.FloatField()# mJy/beam
-    flux_peak_isl_ratio = models.FloatField(
-        help_text=(
-            'Ratio of the component peak flux to the total'
-            ' island peak flux.'
-        )
-    )
-    chi_squared_fit = models.FloatField(
-        db_column='chi2_fit',
-        help_text='Chi-squared of the Guassian fit to the source.'
-    )
-    spectral_index = models.FloatField(
-        db_column='spectr_idx',
-        help_text='In-band Selavy spectral index.'
-    )
-    spectral_index_from_TT = models.BooleanField(
-        default=False,
-        db_column='spectr_idx_tt',
-        help_text=(
-            'True/False if the spectral index came from the taylor '
-            'term.'
-        )
-    )
-
-    local_rms = models.FloatField(
-        help_text='Local rms in mJy from Selavy.'
-    )# mJy/beam
-
-    snr = models.FloatField(
-        help_text='Signal-to-noise ratio of the measurement.'
-    )
-
-    flag_c4 = models.BooleanField(
-        default=False,
-        help_text='Fit flag from Selavy.'
-    )
-
-    compactness = models.FloatField(
-        help_text='Int flux over peak flux.'
-    )
-
-    has_siblings = models.BooleanField(
-        default=False,
-        help_text='True if the fit comes from an island that has more than 1 component.'
-    )
-    component_id = models.CharField(
-        max_length=64,
-        help_text=(
-            'The ID of the component from which the source comes from.'
-        )
-    )
-    island_id = models.CharField(
-        max_length=64,
-        help_text=(
-            'The ID of the island from which the source comes from.'
-        )
-    )
-
-    forced = models.BooleanField(
-        default=False,
-        help_text='True: the measurement is forced extracted.'
-    )
-
-    objects = MeasurementQuerySet.as_manager()
-
-    class Meta:
-        ordering = ['ra']
-
-    def __str__(self):
-        return self.name
+            m_peak = calculate_m_metric(meas_a.flux_peak, meas_b.flux_peak)
+            measurement_pairs.append(
+                MeasurementPair(self.id, meas_a.id, meas_b.id, vs_peak, m_peak, vs_int, m_int)
+            )
+        return measurement_pairs
 
 
 class Association(models.Model):
@@ -720,6 +740,28 @@ class Association(models.Model):
         )
 
 
+class RelatedSource(models.Model):
+    '''
+    Association table for the many to many Source relationship with itself
+    Django doc
+    https://docs.djangoproject.com/en/3.1/ref/models/fields/#django.db.models.ManyToManyField.through
+    '''
+    from_source = models.ForeignKey(Source, on_delete=models.CASCADE)
+    to_source = models.ForeignKey(
+        Source,
+        on_delete=models.CASCADE,
+        related_name='related_sources'
+    )
+
+    class Meta:
+        constraints = [
+            models.UniqueConstraint(
+                name='%(app_label)s_%(class)s_unique_pair',
+                fields=['from_source', 'to_source']
+            )
+        ]
+
+
 class SourceFav(models.Model):
     user = models.ForeignKey(User, on_delete=models.CASCADE)
     source = models.ForeignKey(Source, on_delete=models.CASCADE)
@@ -730,35 +772,3 @@ class SourceFav(models.Model):
         blank=True,
         help_text='Why did you include this as favourite'
     )
-
-
-class MeasurementPair(models.Model):
-    """Links two Measurement objects from the same Source and stores two variability metrics
-    for peak and integrated fluxes:
-        - `vs_peak` and `vs_int` is the variability t-statistic. e.g. if Vs is 4.3, then
-            the source is considered variable to a 95% CI.
-        - `m_peak` and `m_int` is the modulation index, related to fractional variability.
-    See Section 5 of Mooley et al. (2016) for details, DOI: 10.3847/0004-637X/818/2/105.
-    """
-    source = models.ForeignKey(Source, on_delete=models.CASCADE)
-    measurement_a = models.ForeignKey(
-        Measurement, related_name="measurement_pairs_a", on_delete=models.CASCADE
-    )
-    measurement_b = models.ForeignKey(
-        Measurement, related_name="measurement_pairs_b", on_delete=models.CASCADE
-    )
-    vs_peak = models.FloatField(help_text="Variability metric: t-statistic for peak fluxes.")
-    m_peak = models.FloatField(help_text="Variability metric: modulation index for peak fluxes.")
-    vs_int = models.FloatField(help_text="Variability metric: t-statistic for integrated fluxes.")
-    m_int = models.FloatField(help_text="Variability metric: modulation index for integrated fluxes.")
-
-    class Meta:
-        constraints = [
-            models.UniqueConstraint(
-                fields=["source", "measurement_a", "measurement_b"],
-                name="%(app_label)s_%(class)s_unique_pair"
-            )
-        ]
-
-    def __str__(self):
-        return f"({self.measurement_a}, {self.measurement_b})"
